Update unit dp methods and related object attributes
To address the use case where a queryString includes the unitdp=4 param we need to change objects with attributes that support more than two decimal points to to change the number datatype from double to float.

Added missing unitdp to methods ...

Added to the following ..

BankTransactions

getBankTransaction
createBankTransactions
updateBankTransaction
updateOrCreate
CreditNotes

getCreditNotes
createCreditNotes
updateCreditNote
updateOrCreateCreditNotes
Invoices

getInvoice
createInvoices
updateInvoice
updateOrCreateInvoices
Items

getItem
createItems
updateItem
updateOrCreateItems
Receipts

getReceipts
getReceipt
createReceipt
updateReceipt
Objects that use LineItems are impacted

BankTransations
CreditNotes
Invoices
Receipts
the two properties that changed from double to float

Quantity
UnitAmount
The Purchase object property changed from double to float

UnitPrice
The Item object has the property changed from double to float

QuantityOnHand
Added Reference attribute to Prepayment object
Added VOIDED status to Receipt object
Added DELETED status to RepeatingInvoice object

+++ lib/Models/Accounting/Item.php
@@ -68,7 +68,7 @@ class Item implements ModelInterface, ArrayAccess
         'sales_details' => '\XeroAPI\XeroPHP\Models\Accounting\Purchase',
         'is_tracked_as_inventory' => 'bool',
         'total_cost_pool' => 'double',
-        'quantity_on_hand' => 'double',
+        'quantity_on_hand' => 'float',
         'updated_date_utc' => '\DateTime',
         'item_id' => 'string',
         'status_attribute_string' => 'string',
@@ -92,7 +92,7 @@ class Item implements ModelInterface, ArrayAccess
         'sales_details' => null,
         'is_tracked_as_inventory' => null,
         'total_cost_pool' => 'double',
-        'quantity_on_hand' => 'double',
+        'quantity_on_hand' => 'float',
         'updated_date_utc' => 'date-time',
         'item_id' => 'uuid',
         'status_attribute_string' => null,
@@ -596,7 +596,7 @@ class Item implements ModelInterface, ArrayAccess
     /**
      * Gets quantity_on_hand
      *
-     * @return double|null
+     * @return float|null
      */
     public function getQuantityOnHand()
     {
@@ -606,7 +606,7 @@ class Item implements ModelInterface, ArrayAccess
     /**
      * Sets quantity_on_hand
      *
-     * @param double|null $quantity_on_hand The quantity of the item on hand
+     * @param float|null $quantity_on_hand The quantity of the item on hand
      *
      * @return $this
      */
